LYMON-692 refactor(audit): centralize audit diff logic into utility function and update related handlers to reduce duplicated code
